% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNABECNet}
\alias{getNABECNet}
\title{getNABECNet - Accessing a network object directly}
\usage{
getNABECNet(tissue = getNABECTissues())
}
\arguments{
\item{tissue}{One of the labels that can be obtained by calling getGTExTissues() method
to refer to a specific network within the package}
}
\value{
The RDS full file name to the network
}
\description{
getNABECNet - Accessing a network object directly
}
